backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
redis==5.0.7
httpx==0.27.2
anthropic==0.34.2
